+++ Basalt/Common/Utils/BasaltConstants.cs
@@ -6,5 +6,6 @@
 		public const string UpdateEventKey = "update";
 		public const string PhysicsUpdateEventKey = "physics_update";
 		public const string RenderEventKey = "render";
+		public const string UiRenderEventKey = "ui_render";
 	}
 }