+++ Basalt/Types/UIPivot.cs
@@ -1,0 +1,15 @@
+﻿namespace Basalt.Types
+{
+	public enum UIPivot
+	{
+		TopLeft,
+		TopCenter,
+		TopRight,
+		MiddleLeft,
+		MiddleCenter,
+		MiddleRight,
+		BottomLeft,
+		BottomCenter,
+		BottomRight
+	}
+}